requirements.txt
python-dotenv
requests
dataclasses-json
mysql-connector-python
schedule
tabulate